64 bit fixes

+++ src/cgame/cg_syscalls.c
@@ -30,13 +30,13 @@ If you have questions concerning this license or the applicable additional terms
 // cg_syscalls.asm is included instead when building a qvm
 #include "cg_local.h"
 
-static int ( QDECL * syscall )( int arg, ... ) = ( int ( QDECL * )( int, ... ) ) - 1;
+static intptr_t ( QDECL * syscall )( int arg, ... ) = ( int ( QDECL * )( int, ... ) ) - 1;
 
 // TTimo: guarding
 #if defined( __MACOS__ )
 #pragma export on
 #endif
-void dllEntry( int ( QDECL  *syscallptr )( int arg,... ) ) {
+void dllEntry( intptr_t ( QDECL  *syscallptr )( int arg,... ) ) {
 #if defined( __MACOS__ )
 #pragma export off
 #endif
